// src/app/shared/enums/categories.ts
export enum Races {
  WHITE = "white",
  COLOURED = "coloured",
  INDIAN = "indian",
  BLACK = "black"
}

export enum CategoryBreakdownTitles {
  YEAR = "Year",
  FUNDS = "Total funds allocated",
  STUDENTS = "Total students funded",
  PERCENTAGES = "Total percentage of funding per race",
  PERCENTAGES_HEADING = "Total percentage of funding per race",
  TOTAL = "Total",
  FILTER_DEFAULT = "All",
  STUDENTS_LABEL = "Students",
  USED = "Amount used",
  FUND_ALLOCATION = "Fund allocation per university",
  CHART_TYPE = "bar"
}

export enum Colours {
  SHADE_1 = "#584ee4",
  SHADE_2 = "#9b94ff",
  SHADE_3 = "#39339B",
  SHADE_4 = "#25225E",
  DEFAULT = "#FFFFFF"
}